+++ __tests__/controllers/modules/auth.test.js
@@ -17,6 +17,11 @@ beforeEach(() => {
 
 // test
 describe('Modules: Auth', () => {
+  test('Middleware: usePassport', () => {
+    // no testing items
+  });
+
+
   test('Middleware: isSignedIn', () => {
     const _isSignedIn = isSignedIn[isSignedIn.length - 1];
     req.session.user = {};