+++ controllers/routers/auth.js
@@ -8,11 +8,11 @@ const { ClientException } = require('../utilities/')._U_.error;
 const auth = {};
 
 auth.signup = {
-  GET: function account_signup_GET(req, res, next) {
+  GET: function auth_signup_GET(req, res, next) {
     if (req.isAuthenticated()) return res.redirect('/home');
     return next();
   },
-  POST: [_M_.isValidPasswordReset, async function account_signup_POST(req, res) {
+  POST: [_M_.isValidPasswordReset, async function auth_signup_POST(req, res) {
     const newUser = await UsersModel.register(new UsersModel(req.body), req.body.password.new);
     req.logIn(newUser, err => {
       if (err) throw err;
@@ -24,12 +24,12 @@ auth.signup = {
 
 
 auth.signin = {
-  GET: function account_signin_GET(req, res, next) {
+  GET: function auth_signin_GET(req, res, next) {
     if (req.isAuthenticated()) return res.redirect('/home');
     if (res.locals.$$VIEW.flash.action.includes('retry')) req.flash('action', 'retry');                                 // note: // preserve `returnTo` if client not visit any other pages
     return next();
   },
-  POST: function account_signin_POST(req, res, next) {
+  POST: function auth_signin_POST(req, res, next) {
     if (req.isAuthenticated()) return res.redirect('/home');
     return passport.authenticate('local', (authErr, authUser) => {
       if (authErr) return next(authErr);
@@ -48,7 +48,7 @@ auth.signin = {
 
 
 auth.signout = {
-  GET: function account_signout_GET(req, res) {
+  GET: function auth_signout_GET(req, res) {
     if (req.isAuthenticated()) {
       req.logout();
       req.flash('info', 'See you next time!');