+++ routers/auth.js
@@ -1,4 +1,3 @@
-/* eslint-disable key-spacing */
 const { usePassport } = require('../controllers/modules/')._M_;
 const { ClientException } = require('../controllers/utilities/')._U_.error;
 const { Device } = require('../controllers/engines/router');
@@ -8,18 +7,24 @@ const { auth } = require('../controllers/routers/');
 // device
 const authRouter = new Device([
   {
-    route:        '/signup',
-    controller:   auth.signup,
-    setting:      { title: 'Sign Up', template: './__root__/account/signup' },
+    route: '/signup',
+    controller: auth.signup,
+    setting: {
+      title: 'Sign Up',
+      template: './__root__/account/signup',
+    },
   },
   {
-    route:        '/signin',
-    controller:   auth.signin,
-    setting:      { title: 'Sign In', template: './__root__/account/signin' },
+    route: '/signin',
+    controller: auth.signin,
+    setting: {
+      title: 'Sign In',
+      template: './__root__/account/signin',
+    },
   },
   {
-    route:        '/signout',
-    controller:   auth.signout,
+    route: '/signout',
+    controller: auth.signout,
   },
 ]);