+++ routers/blog.js
@@ -1,31 +1,60 @@
-/* eslint-disable key-spacing */
 const { Device } = require('../controllers/engines/router');
-const { editor, blog } = require('../controllers/routers/');
+const { blog } = require('../controllers/routers/');
 
 
 // device
 const blogRouter = new Device([
   {
-    route:        ['/edit', '/new'],
-    controller:   editor.new,
-    setting:      { template: './__root__/editor', title: 'New post', authentication: true },
+    route: ['/edit', '/new'],
+    controller: blog.edit,
+    permission: {
+      access: ['editor'],                                                                                               // todo: use Device.permission.ROLE_EDITOR
+      change: ['editor'],
+    },
+    setting: {
+      method: 'GET',
+      template: './__root__/editor',
+      title: 'New post',
+    },
   },
   {
-    route:        /^\/([a-f\d]{24})\/?\/edit\/?$/i,
-    alias:        '/:canonical/edit',
-    controller:   editor.post,
-    setting:      { template: './__root__/editor', title: 'Edit post', authorization: true },
+    alias: '/:canonical/edit',
+    route: /^\/([a-f\d]{24})\/?\/edit\/?$/i,
+    controller: blog.edit,
+    permission: {
+      access: ['owner'],
+      change: ['owner'],
+    },
+    setting: {
+      template: './__root__/editor',
+      title: 'Edit post',
+    },
   },
   {
-    route:        /^\/(?![a-f\d]{24}\/?$)([^\/]+)\/?$/i,
-    alias:        '/:_id',
-    controller:   blog.post,
-    setting:      { template: './theme/blog/posts', renderer: Device.renderer.VIEW_POSTS_SINGLE, servingAPI: true },
+    alias: '/:canonical',
+    route: /^\/([a-f\d]{24})\/?$/i,
+    controller: blog.post,
+    permission: {
+      access: ['public'],
+      change: ['owner'],
+    },
+    setting: {
+      template: './theme/blog/posts',
+      servingAPI: true,
+    },
   },
   {
-    route:        '/',
-    controller:   blog.list,
-    setting:      { template: './theme/blog/', renderer: Device.renderer.VIEW_POSTS_MULTIPLE, servingAPI: true },
+    route: '/',
+    controller: blog.list,
+    permission: {
+      access: ['public'],
+      change: ['editor'],
+    },
+    setting: {
+      template: './theme/blog/',
+      renderer: Device.renderer.VIEW_POSTS_MULTIPLE,
+      servingAPI: true,
+    },
   },
 ]);