+++ services/error.js
@@ -5,7 +5,10 @@ const { exceptionHandler } = require('../controllers/handlers/error');
 const errorHandlingAgent = app => {
   app.use('*', (req, res, next) => next(new HttpException(404)));
   app.use(exceptionHandler);
-  app.use((err, req, res, next) => res.sendStatus(500));
+  app.use((err, req, res, next) => {
+    console.log(err);
+    res.sendStatus(500);
+  });
 };